refactor for all utils tests add test for run_command
