+++ q2_amr/card/mags.py
@@ -7,7 +7,8 @@ import pandas as pd
 from q2_types.per_sample_sequences import MultiMAGSequencesDirFmt
 
 from q2_amr.card.types import CARDAnnotationDirectoryFormat, CARDDatabaseDirectoryFormat
-from q2_amr.card.utils import create_count_table, load_card_db, read_in_txt, run_command
+from q2_amr.card.utils import load_card_db
+from q2_amr.utils.utils import create_count_table, read_in_txt, run_command
 
 
 def annotate_mags_card(
