+++ q2_amr/card/reads.py
@@ -18,7 +18,8 @@ from q2_amr.card.types import (
     CARDDatabaseDirectoryFormat,
     CARDGeneAnnotationDirectoryFormat,
 )
-from q2_amr.card.utils import create_count_table, load_card_db, read_in_txt, run_command
+from q2_amr.card.utils import load_card_db
+from q2_amr.utils.utils import create_count_table, read_in_txt, run_command
 
 
 def annotate_reads_card(
