+++ q2_amr/utils/tests/data/output.allele_mapping_data.txt
@@ -1,0 +1,5 @@
+Reference Sequence	ARO Term	ARO Accession	Reference Model Type	Reference DB	Reference Allele Source	Resistomes & Variants: Observed in Genome(s)	Resistomes & Variants: Observed in Plasmid(s)	Resistomes & Variants: Observed Pathogen(s)	Completely Mapped Reads	Mapped Reads with Flanking Sequence	All Mapped Reads	Percent Coverage	Length Coverage (bp)	Average MAPQ (Completely Mapped Reads)	Mate Pair Linkage	Reference Length	AMR Gene Family	Drug Class	Resistance Mechanism
+ARO:3000796|ID:121|Name:mdtF|NCBI:U00096.1	mdtF	3000796	protein homolog model	CARD	CARD curation	no data	no data	Escherichia coli	1	0	1	8.09	252	193.00	*	3114	resistance-nodulation-cell division (RND) antibiotic efflux pump	macrolide antibiotic; fluoroquinolone antibiotic; penam	antibiotic efflux
+ARO:3000815|ID:154|Name:mgrA|NCBI:BA000018.3	mgrA	3000815	protein homolog model	CARD	CARD curation	no data	no data	Staphylococcus aureus	1	0	1	19.59	87	172.00	*	444	ATP-binding cassette (ABC) antibiotic efflux pump; major facilitator superfamily (MFS) antibiotic efflux pump	fluoroquinolone antibiotic; cephalosporin; penam; tetracycline antibiotic; peptide antibiotic; disinfecting agents and antiseptics	antibiotic efflux
+ARO:3000805|ID:172|Name:OprN|NCBI:AE004091.2	OprN	3000805	protein homolog model	CARD	CARD curation	no data	no data	Pseudomonas aeruginosa	1	0	1	17.76	252	193.00	*	1419	resistance-nodulation-cell division (RND) antibiotic efflux pump	fluoroquinolone antibiotic; diaminopyrimidine antibiotic; phenicol antibiotic	antibiotic efflux
+ARO:3000026|ID:377|Name:mepA|NCBI:AY661734.1	mepA	3000026	protein homolog model	CARD	CARD curation	no data	no data	Staphylococcus aureus	1	0	1	17.70	240	190.50	*	1356	multidrug and toxic compound extrusion (MATE) transporter	glycylcycline; tetracycline antibiotic	antibiotic efflux
