+++ q2_amr/utils/tests/data/output.gene_mapping_data.txt
@@ -1,0 +1,5 @@
+ARO Term	ARO Accession	Reference Model Type	Reference DB	Alleles with Mapped Reads	Reference Allele(s) Identity to CARD Reference Protein (%)	Resistomes & Variants: Observed in Genome(s)	Resistomes & Variants: Observed in Plasmid(s)	Resistomes & Variants: Observed Pathogen(s)	Completely Mapped Reads	Mapped Reads with Flanking Sequence	All Mapped Reads	Average Percent Coverage	Average Length Coverage (bp)	Average MAPQ (Completely Mapped Reads)	Number of Mapped Baits	Number of Mapped Baits with Reads	Average Number of reads per Bait	Number of reads per Bait Coefficient of Variation (%)	Number of reads mapping to baits and mapping to complete gene	Number of reads mapping to baits and mapping to complete gene (%)	Mate Pair Linkage (# reads)	Reference Length	AMR Gene Family	Drug Class	Resistance Mechanism
+mdtF	3000796	protein homolog model	CARD	1	100.0	no data	no data	Escherichia coli	1.00	0.00	1	8.09	252.00	193.00	0	0	0	0	N/A	N/A		3114	resistance-nodulation-cell division (RND) antibiotic efflux pump	macrolide antibiotic; fluoroquinolone antibiotic; penam	antibiotic efflux
+mgrA	3000815	protein homolog model	CARD	1	100.0	no data	no data	Staphylococcus aureus	1.00	0.00	1	19.59	87.00	172.00	0	0	0	0	N/A	N/A		444	ATP-binding cassette (ABC) antibiotic efflux pump; major facilitator superfamily (MFS) antibiotic efflux pump	fluoroquinolone antibiotic; cephalosporin; penam; tetracycline antibiotic; peptide antibiotic; disinfecting agents and antiseptics	antibiotic efflux
+OprN	3000805	protein homolog model	CARD	1	100.0	no data	no data	Pseudomonas aeruginosa	1.00	0.00	1	17.76	252.00	193.00	0	0	0	0	N/A	N/A		1419	resistance-nodulation-cell division (RND) antibiotic efflux pump	fluoroquinolone antibiotic; diaminopyrimidine antibiotic; phenicol antibiotic	antibiotic efflux
+mepA	3000026	protein homolog model	CARD	1	100.0	no data	no data	Staphylococcus aureus	1.00	0.00	1	17.70	240.00	190.50	0	0	0	0	N/A	N/A		1356	multidrug and toxic compound extrusion (MATE) transporter	glycylcycline; tetracycline antibiotic	antibiotic efflux
